+++ src/Token.php
@@ -2,6 +2,7 @@
 namespace Pyncer\Utility;
 
 use Pyncer\Utility\EqualsInterface;
+use Stringable;
 
 use function base64_encode;
 use function hash_equals;
@@ -12,7 +13,7 @@ use function substr;
 
 class Token implements EqualsInterface
 {
-    private $token;
+    private string $token;
 
     public function __construct(?string $token = null)
     {
@@ -35,7 +36,17 @@ class Token implements EqualsInterface
 
     public function equals(mixed $value): bool
     {
-        return hash_equals($this->token, $token);
+        if ($value instanceof Token) {
+            $value = $value->getValue();
+        } elseif ($value instanceof Stringable) {
+            $value = strval($value);
+        }
+
+        if (!is_string($value)) {
+            return false;
+        }
+
+        return hash_equals($this->token, $value);
     }
 
     private function generateToken(): string