+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Change Log
 
+## 1.0.1 - 2023-09-16
+
+### Added
+
+- Added PHPStan static analysis.
+
+### Changed
+
+- The ParamsInterface::getStr() and ParamsInterface::setStr() functions have been deprecated in favor of ParamsInterface::getString and ParamsInterface::setString().
+
 ## 1.0.0 - 2022-12-28
 
 Initial release.